requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-chroma
openai
python-dotenv
pypdf
chromadb
streamlit

